% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_off_selecting.R
\name{cut_off_selecting}
\alias{cut_off_selecting}
\title{selecting cutofff by proc output}
\usage{
cut_off_selecting(roc, cutoff_type = "youden", sen = 0.9, spe = 0.9)
}
\arguments{
\item{roc}{proc output}

\item{sen}{let the sensitivities is at least one value}

\item{spe}{let the specificities is at least one value}

\item{cutoff}{one value for divide data into two group}
}
\value{
cutoff value
}
\description{
selecting cutofff by proc output
}
\examples{
library(Clabomic)
library(pROC)
colnames(iris)[ncol(iris)] <- "group"
roc <- roc(iris$group, iris$Sepal.Length)
cutoff <- cut_off_selecting(roc)
}
