% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_batch_feature_sel.R
\name{do_batch_feature_sel}
\alias{do_batch_feature_sel}
\title{The last column is used as grouping, and the other columns are subjected to feature project}
\usage{
do_batch_feature_sel(mat)
}
\arguments{
\item{mat}{a data frame,the last column is 'group', and the content is binary classification text}
}
\value{
features weights with three tools
}
\description{
The last column is used as grouping, and the other columns are subjected to feature project
}
\examples{
library(Clabomic)
colnames(iris)[ncol(iris)] <- "group"
res <- do_batch_feature_sel(iris)
}
