% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_batch_roc.R
\name{do_batch_roc}
\alias{do_batch_roc}
\title{The last column is used as grouping, and the other columns are subjected to roc test}
\usage{
do_batch_roc(dat)
}
\arguments{
\item{dat}{a data frame,the last column is 'group', and the content is binary classification text}
}
\value{
roc test results
}
\description{
The last column is used as grouping, and the other columns are subjected to roc test
}
\examples{
library(Clabomic)
colnames(iris)[ncol(iris)] <- "group"
res <- do_batch_roc(iris)
}
