% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_batch_Wilcoxon.R
\name{do_batch_Wilcoxon}
\alias{do_batch_Wilcoxon}
\title{The last column is used as grouping, and the other columns are subjected to Wilcoxon test}
\usage{
do_batch_Wilcoxon(mat)
}
\arguments{
\item{mat}{a data frame,the last column is 'group', and the content is binary classification text}
}
\value{
Wilcoxon test results
}
\description{
The last column is used as grouping, and the other columns are subjected to Wilcoxon test
}
\examples{
library(Clabomic)
colnames(iris)[ncol(iris)] <- "group"
iris_batch_wilcoxon_res <- do_batch_Wilcoxon(iris)
}
