% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ga_based_svm.R
\name{do_ga_based_svm}
\alias{do_ga_based_svm}
\title{Optimization of svm high-dimensional parameters based on genetic algorithm}
\usage{
do_ga_based_svm(dat, positive_col = "malignant", maxiter = 100)
}
\arguments{
\item{dat}{data frame with several numeric features and one Categorical variables at last column}

\item{positive_col}{the name of positive Categorical variable at last column}

\item{maxiter}{max generation}
}
\value{
fig and value
}
\description{
Optimization of svm high-dimensional parameters based on genetic algorithm
}
\examples{
colnames(iris)[ncol(iris)] <- "group"
iris$group <- ifelse(iris$group == "setosa", "pos", "nag")
res <- do_ga_based_svm(iris, positive_col = "pos")
}
