% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_muti_roc.R
\name{do_muti_roc}
\alias{do_muti_roc}
\title{Constuct several roc lines in one fig}
\usage{
do_muti_roc(roc.list)
}
\arguments{
\item{roc.list}{}
}
\value{

}
\description{
Constuct several roc lines in one fig
}
\examples{
library(Clabomic)
library(pROC)
colnames(iris)[ncol(iris)] <- "group"
roc.list <- list(Train = roc(iris$group, iris$Sepal.Length), Test = roc(iris$group, iris$Sepal.Width))
do_muti_roc(roc.list)
}
