% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_svm.R
\name{get_best_svm}
\alias{get_best_svm}
\title{To get bset svm model}
\usage{
get_best_svm(Train, res)
}
\arguments{
\item{Train}{training data}

\item{res}{result from do_ga_based_svm.R}
}
\value{
best_model
}
\description{
To get bset svm model
}
\examples{
library(Clabomic)
colnames(iris)[ncol(iris)] <- "group"
iris$group <- ifelse(iris$group == "setosa", "pos", "nag")
res <- do_ga_based_svm(iris, positive_col = "pos")
iris$group=as.numeric(ifelse(iris$group == "pos", 1, 0))
best_model <- get_best_svm(iris,res)
}
