% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_importance.R
\name{svm_importance}
\alias{svm_importance}
\title{svm_importance evaluate}
\usage{
svm_importance(best_model)
}
\arguments{
\item{best_model}{best_model by cost and gamma by do_ga_based_svm}
}
\value{
svm_importance data frame
}
\description{
svm_importance evaluate
}
\examples{
library(Clabomic)
colnames(iris)[ncol(iris)] <- "group"
iris$group <- ifelse(iris$group == "setosa", "pos", "nag")
res <- do_ga_based_svm(iris, positive_col = "pos")
best_model <- svm(group ~ ., data = Train, gamma = signif(res[[2]]$solution[1], 3), cost = signif(res[[2]]$solution[2], 3))
importance <- svm_importance(best_model)
}
